Fix/internet access oth fixes (#2)
* 1. internet access checker before request
2. texts in btn changed
3. next btn in keyboard fixed
4. kotlin version upgraded

* 1. fixed parser *receiverUUID

+++ app/src/main/java/kz/das/dasaccounting/ui/drivers/transfer/TransferConfirmVM.kt
@@ -9,6 +9,7 @@ import kz.das.dasaccounting.core.ui.view_model.BaseVM
 import kz.das.dasaccounting.data.entities.common.TransferItem
 import kz.das.dasaccounting.domain.DriverInventoryRepository
 import kz.das.dasaccounting.domain.data.drivers.TransportInventory
+import kz.das.dasaccounting.utils.InternetAccess
 import org.koin.core.inject
 
 class TransferConfirmVM: BaseVM() {
@@ -86,25 +87,33 @@ class TransferConfirmVM: BaseVM() {
 
     fun sendInventory() {
         viewModelScope.launch {
-            showLoading()
-            try {
-                transportInventory?.apply {
-                    latitude = getUserLocation().lat
-                    longitude = getUserLocation().long
-                    writeObjectToLog(this.toString(), context)
+            if (InternetAccess.internetCheck(context)) {
+                showLoading()
+                try {
+                    transportInventory?.apply {
+                        latitude = getUserLocation().lat
+                        longitude = getUserLocation().long
+                        writeObjectToLog(this.toString(), context)
 
-                    transportInventoryRepository.sendInventory(this)
-                    transportInventoryRepository.removeItem(this)
+                        transportInventoryRepository.sendInventory(this)
+                        transportInventoryRepository.removeItem(this)
+                    }
+                    isTransportInventorySentLV.postValue(true)
+                } catch (t: Throwable) {
+                    transportInventory?.let {
+                        transportInventoryRepository.saveAwaitSentInventory(it)
+                    }
+                    isOnAwaitLV.postValue(true)
+                } finally {
+                    deleteSavedInventory()
+                    hideLoading()
                 }
-                isTransportInventorySentLV.postValue(true)
-            } catch (t: Throwable) {
+            } else {
                 transportInventory?.let {
                     transportInventoryRepository.saveAwaitSentInventory(it)
                 }
                 isOnAwaitLV.postValue(true)
-            } finally {
                 deleteSavedInventory()
-                hideLoading()
             }
         }
     }
